Change loop type

+++ decoder/LAVAudio/parser/parser.h
@@ -22,10 +22,10 @@
 static inline const uint8_t *find_marker32_position(const uint8_t *pBuffer, size_t uBufSize, uint32_t marker)
 {
   uint32_t state = 0;
-  for (unsigned i = 0; i < uBufSize; ++i) {
+  for (size_t i = 0; i < uBufSize; ++i) {
     state = (state << 8) | pBuffer[i];
     if (state == marker) {
-      return pBuffer + (i-3);
+      return pBuffer + (i - 3);
     }
   }
